requirements.txt
anyio==3.7.0
asgiref==3.7.2
async-timeout==4.0.2
backports.zoneinfo==0.2.1
bidict==0.22.1
blinker==1.6.2
channels==4.0.0
channels-redis==4.1.0
click==8.1.3
Django==4.2.1
django-cors-headers==4.0.0
djangorestframework==3.14.0
exceptiongroup==1.1.1
h11==0.14.0
httptools==0.5.0
idna==3.4
importlib-metadata==6.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
msgpack==1.0.5
psycopg2-binary==2.9.3
python-dotenv==1.0.0
python-engineio==4.4.1
python-socketio==5.8.0
pytz==2023.3
PyYAML==6.0
redis==4.5.5
sniffio==1.3.0
sqlparse==0.4.4
typing-extensions==4.6.2
uvicorn==0.22.0
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.3
Werkzeug==2.3.4
zipp==3.15.0